requirements.txt
absl-py==1.0.0
nltk=3.6.7
astunparse==1.6.3
brotli==1.0.9
cachetools
charset-normalizer==2.0.12
flatbuffers==2.0
gast
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py
idna==3.3
importlib-metadata
keras==2.7.0
keras-preprocessing==1.1.2
keras-self-attention==0.51.0
keras-tuner==1.1.2
kt-legacy==1.0.4
libclang==13.0.0
markdown==3.3.6
oauthlib==3.2.0
opt-einsum==3.3.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.1
install requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing-extensions==4.1.1
urllib3==1.26.8
werkzeug==2.0.3
wrapt==1.14.0
zipp
fasttext
modA


